requirements.txt
Django==1.8.1
django-tastypie==0.12.2-dev
